update main.ts to reuse snippets from Custom-Krnl-Monaco code
This project is supposed to be a rewrite of Custom-Krnl-Monaco.
However, I rebranded it because Krnl is dead and this project isn't going to be just for Krnl anymore.
Also, you can't even use Krnl because they took down all of their APIs and servers (probably including the keysystem API and other stuff the DLL relied on).