add backup data for auto-complete offline
The idea is that this will be bundled with the project so that if you clone the code and work offline, even if you don't have the autocomplete data in localStorage, you should be able to get completion.
The only problem with backup measure #1 is that the storage is limited. Maybe we should have a server and it'd store the autocomplete metadata?
There should be 2 backup measures for offline editing:
1. Last API-Dump and Autocomplete metadata dump (this is the ideal option because it should be more up to date)
2. Backup files bundled with this project (last resort, it may be less up to date, especially if the current revision being used is out of date)