Fix CI errors
drops support for browsers that don't have support for ESM or TLA (like safari 11 and chrome 88), I don't care that you still want to use your 69420 year old browser, just upgrade it.
It's essentially the same thing to me as asking for me to support Firefox 1.
I can support newer versions, but not the old ones.